Get ready to pass MATCH information to filter method
I want to use MATCH against three columns - timestamp, cwd, and entry
- and the way I want MATCH to work will differ among those three columns.
I can't get away with using find_function to return an implementation
for MATCH, since if I return an implementation for MATCH via find_function, that
function won't be able to differentiate which of those columns it's
being used against (I *might* be able to use sqlite3_set_auxdata for
this, but since I'm marshalling all SQLite values into and out of Lua
values, I don't have a good interface for doing this, nor do I have any
ideas for what such an interface should look like), and so I would have
to rely on the syntax of the match expression on the right-side hand to
guess which column is on the left-hand side, and I don't care for that
limitation.

However, I only ever want to use MATCH with a literal right-hand side
anyway, and that information gets passed to best_index - what's more,
best_index can tell SQLite to ignore that condition in the SQL via
omit, so I can optimize away calls to MATCH, removing my need for
find_function at all!

I still need to get the filter method to set up a cursor that does the
right filtering based on MATCH expressions in the WHERE clause, though -
best_index can only communicate the name of the index to use, plus which
constraints' arguments to pass along to filter.  AFAICT you can't have
gaps in the args array that ends up getting passed to filter, so I can't
do something nice like say "the first argument is NULL or the operand
for 'timestamp MATCH X'".

So what I did is encode the column names and args positions for
conditions using MATCH into a virtual index name - the filter method
will decode this and figure out what extra stuff it needs to add to the
WHERE clause to do its job.