Add keybinding to mark a browser session as noteworthy
This is so that when I'm using the tool, if I run into behavior that's
weird that I don't feel like digging into in the moment, I can "flag" it
for later analysis.  Right now that's just a matter of emitting a
special log message - maybe later it'll be something more like dumping
the database and the bubbletea message history to a directory or
something.